rename: ChoiceNotAllowedException Package 변경
