+++ src/main/java/com/zipline/global/exception/custom/ChoiceNotAllowedException.java
@@ -1,8 +1,6 @@
-package com.zipline.global.exception.custom.customer;
+package com.zipline.global.exception.custom;
 
 import org.springframework.http.HttpStatus;
-
-import com.zipline.global.exception.custom.BaseException;
 
 public class ChoiceNotAllowedException extends BaseException {
 	public ChoiceNotAllowedException(String message, HttpStatus status) {
